refactor: GAトラッキングデータ送信のファイルをgaディレクトリ内に移動
